Fix lint issues in the friend list and search screens
