--- src/components/Social/SearchResult.tsx
@@ -1,171 +1,0 @@
-import React from 'react';
-import {StyleSheet, View} from 'react-native';
-import type {Database} from 'firebase/database';
-import type {FirebaseStorage} from 'firebase/storage';
-import CONST from '@src/CONST';
-import ProfileImage from '@components/ProfileImage';
-import type {FriendRequestStatus, Profile} from '@src/types/onyx';
-import * as ErrorUtils from '@libs/ErrorUtils';
-import Button from '@components/Button';
-import useThemeStyles from '@hooks/useThemeStyles';
-import Text from '@components/Text';
-import {acceptFriendRequest, sendFriendRequest} from '@src/database/friends';
-import ERRORS from '@src/ERRORS';
-import useLocalize from '@hooks/useLocalize';
-
-type SendFriendRequestButtonProps = {
-  db: Database;
-  userFrom: string;
-  userTo: string;
-  requestStatus: FriendRequestStatus | undefined;
-  alreadyAFriend: boolean;
-};
-
-const SendFriendRequestButton: React.FC<SendFriendRequestButtonProps> = ({
-  db,
-  userFrom,
-  userTo,
-  requestStatus,
-  alreadyAFriend,
-}) => {
-  const [isLoading, setIsLoading] = React.useState<boolean>(false);
-  const {translate} = useLocalize();
-  const styles = useThemeStyles();
-
-  const handleSendRequestPress = async (
-    db: Database,
-    userFrom: string,
-    userTo: string,
-    setIsLoading: React.Dispatch<React.SetStateAction<boolean>>,
-  ): Promise<void> => {
-    try {
-      setIsLoading(true);
-      await sendFriendRequest(db, userFrom, userTo);
-      setIsLoading(false);
-    } catch (error) {
-      ErrorUtils.raiseAppError(ERRORS.USER.FRIEND_REQUEST_SEND_FAILED, error);
-    }
-  };
-
-  const handleAcceptFriendRequestPress = async (
-    db: Database,
-    userFrom: string,
-    userTo: string,
-    setIsLoading: React.Dispatch<React.SetStateAction<boolean>>,
-  ): Promise<void> => {
-    try {
-      setIsLoading(true);
-      await acceptFriendRequest(db, userFrom, userTo);
-      setIsLoading(false);
-    } catch (error) {
-      ErrorUtils.raiseAppError(ERRORS.USER.FRIEND_REQUEST_SEND_FAILED, error);
-    }
-  };
-
-  return (
-    // Refactor this part using AI later
-    <View style={localStyles.sendFriendRequestContainer}>
-      {userFrom === userTo ? (
-        <Text numberOfLines={1} style={styles.textNormalThemeText}>
-          {translate('searchResult.self')}
-        </Text>
-      ) : alreadyAFriend ? (
-        <Text numberOfLines={1} style={styles.textNormalThemeText}>
-          {translate('searchResult.friend')}
-        </Text>
-      ) : requestStatus === CONST.FRIEND_REQUEST_STATUS.SENT ? (
-        <Text numberOfLines={1} style={styles.textNormalThemeText}>
-          {translate('searchResult.sent')}
-        </Text>
-      ) : requestStatus === CONST.FRIEND_REQUEST_STATUS.RECEIVED ? (
-        <Button
-          add
-          onPress={() => {
-            handleAcceptFriendRequestPress(db, userFrom, userTo, setIsLoading);
-          }}
-          text={translate('searchResult.accept')}
-          isLoading={isLoading}
-        />
-      ) : (
-        <Button
-          add
-          onPress={() => {
-            handleSendRequestPress(db, userFrom, userTo, setIsLoading);
-          }}
-          text={translate('searchResult.add')}
-          isLoading={isLoading}
-        />
-      )}
-    </View>
-  );
-};
-
-type SearchResultProps = {
-  userID: string;
-  userDisplayData: Profile;
-  db: Database;
-  storage: FirebaseStorage;
-  userFrom: string;
-  requestStatus: FriendRequestStatus | undefined;
-  alreadyAFriend: boolean;
-  customButton?: React.ReactNode;
-};
-
-const SearchResult: React.FC<SearchResultProps> = ({
-  userID,
-  userDisplayData,
-  db,
-  storage,
-  userFrom,
-  requestStatus,
-  alreadyAFriend,
-  customButton,
-}) => {
-  const {translate} = useLocalize();
-  const styles = useThemeStyles();
-  return (
-    <View style={styles.userOverviewContainer}>
-      <View style={styles.userOverviewLeftContent}>
-        <ProfileImage
-          key={`${userID}-profile-icon`}
-          storage={storage}
-          userID={userID}
-          downloadPath={userDisplayData?.photo_url}
-          style={localStyles.userProfileImage}
-        />
-        <Text style={[styles.headerText, styles.ml3]}>
-          {userDisplayData?.display_name
-            ? userDisplayData.display_name
-            : translate('common.unknown')}
-        </Text>
-      </View>
-      {customButton || (
-        <SendFriendRequestButton
-          db={db}
-          userFrom={userFrom}
-          userTo={userID}
-          requestStatus={requestStatus}
-          alreadyAFriend={alreadyAFriend}
-        />
-      )}
-    </View>
-  );
-};
-
-// eslint-disable-next-line @typescript-eslint/no-use-before-define
-const localStyles = StyleSheet.create({
-  userProfileImage: {
-    width: 70,
-    height: 70,
-    borderRadius: 35,
-  },
-  sendFriendRequestContainer: {
-    flexShrink: 1,
-    height: '100%',
-    maxHeight: 100,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
-
-export default SearchResult;
